splits functions from test_algorithms.py into fitting modules, also renames and orderes them in a consistent manner
