+++ networkit/test/test_distance.py
@@ -6,61 +6,30 @@ import unittest
 
 import networkit as nk
 
-class TestDistanceAlgorithms(unittest.TestCase):
+class TestDistance(unittest.TestCase):
 
 	def setUp(self):
 		self.L = nk.readGraph("input/looptest1.gml", nk.Format.GML) #without self-loops
 		self.LL = nk.readGraph("input/looptest2.gml", nk.Format.GML) #with self-loops sprinkled in
 
-	def testDistanceDiameter(self):
-		D = nk.distance.Diameter(self.LL, nk.distance.DiameterAlgo.ESTIMATED_RANGE, error = 0.1)
-		D.run()
-		D = nk.distance.Diameter(self.LL, nk.distance.DiameterAlgo.ESTIMATED_SAMPLES, nSamples = 5)
-		D.run()
-		D = nk.distance.Diameter(self.LL, nk.distance.DiameterAlgo.EXACT)
-		D.run()
-
-
-	def testDistanceEccentricity(self):
-		E = nk.distance.Eccentricity()
-		E.getValue(self.LL, 0)
-
-
-	def testDistanceEffectiveDiameter(self):
-		algo = nk.distance.EffectiveDiameter(self.L)
-		algo.run()
-		algo = nk.distance.EffectiveDiameter(self.LL)
-		algo.run()
-
-
-	def testDistanceApproxEffectiveDiameter(self):
-		algo = nk.distance.EffectiveDiameterApproximation(self.L)
-		algo.run()
-		algo = nk.distance.EffectiveDiameterApproximation(self.LL)
-		algo.run()
-
-
-	def testDistanceApproxHopPlot(self):
-		algo = nk.distance.HopPlotApproximation(self.L)
-		algo.run()
-		algo = nk.distance.HopPlotApproximation(self.LL)
-		algo.run()
-
-
-	def testDistanceNeighborhoodFunction(self):
-		algo = nk.distance.NeighborhoodFunction(self.L)
-		algo.run()
-		algo = nk.distance.NeighborhoodFunction(self.LL)
-		algo.run()
-
-
-	def testDistanceApproxNeighborhoodFunction(self):
-		algo = nk.distance.NeighborhoodFunctionApproximation(self.L)
-		algo.run()
-		algo = nk.distance.NeighborhoodFunctionApproximation(self.LL)
-		algo.run()
-
-	def testDistanceAStar(self):
+	def testAsArrayAPSP(self):
+		nk.engineering.setSeed(1, True)
+		random.seed(1)
+		for directed in [True, False]:
+			for weighted in [True, False]:
+				g = nk.generators.ErdosRenyiGenerator(100, 0.15, directed).generate()
+				if weighted:
+					g = nk.graphtools.toWeighted(g)
+					g.forEdges(lambda u, v, ew, eid: g.setWeight(u, v, random.random()))
+				apsp = nk.distance.APSP(g)
+				apsp.run()
+				listDistances = apsp.getDistances()
+				arrayDistances = apsp.getDistances(asarray=True)
+				self.assertIsInstance(listDistances, list)
+				self.assertIsInstance(arrayDistances, np.ndarray)
+				np.testing.assert_allclose(listDistances, arrayDistances)
+	
+	def testAStar(self):
 		# Builds a mesh graph with the given number of rows and columns
 		def buildMesh(rows, cols):
 			G = nk.Graph(rows * cols, False, False)
@@ -129,6 +98,54 @@ class TestDistanceAlgorithms(unittest.TestCase):
 		testMesh(21, 5)
 		testMesh(9, 18)
 		testMesh(7, 1)
+	
+	def testDiameter(self):
+		D = nk.distance.Diameter(self.LL, nk.distance.DiameterAlgo.ESTIMATED_RANGE, error = 0.1)
+		D.run()
+		D = nk.distance.Diameter(self.LL, nk.distance.DiameterAlgo.ESTIMATED_SAMPLES, nSamples = 5)
+		D.run()
+		D = nk.distance.Diameter(self.LL, nk.distance.DiameterAlgo.EXACT)
+		D.run()
+
+
+	def testEccentricity(self):
+		E = nk.distance.Eccentricity()
+		E.getValue(self.LL, 0)
+
+
+	def testEffectiveDiameter(self):
+		algo = nk.distance.EffectiveDiameter(self.L)
+		algo.run()
+		algo = nk.distance.EffectiveDiameter(self.LL)
+		algo.run()
+
+
+	def testApproxEffectiveDiameter(self):
+		algo = nk.distance.EffectiveDiameterApproximation(self.L)
+		algo.run()
+		algo = nk.distance.EffectiveDiameterApproximation(self.LL)
+		algo.run()
+
+
+	def testApproxHopPlot(self):
+		algo = nk.distance.HopPlotApproximation(self.L)
+		algo.run()
+		algo = nk.distance.HopPlotApproximation(self.LL)
+		algo.run()
+
+
+	def testNeighborhoodFunction(self):
+		algo = nk.distance.NeighborhoodFunction(self.L)
+		algo.run()
+		algo = nk.distance.NeighborhoodFunction(self.LL)
+		algo.run()
+
+
+	def testApproxNeighborhoodFunction(self):
+		algo = nk.distance.NeighborhoodFunctionApproximation(self.L)
+		algo.run()
+		algo = nk.distance.NeighborhoodFunctionApproximation(self.LL)
+		algo.run()
 
 	def genERGraphs(self, n = 100, p = 0.15, seed = 42):
 		nk.engineering.setSeed(seed, True)
@@ -141,7 +158,7 @@ class TestDistanceAlgorithms(unittest.TestCase):
 					g.forEdges(lambda u, v, ew, eid: g.setWeight(u, v, random.random()))
 				yield g
 
-	def testDistanceSPSP(self):
+	def testSPSP(self):
 		for g in self.genERGraphs():
 			for nSources in [1, 10, 50]:
 				sources = nk.graphtools.randomNodes(g, nSources)
@@ -164,7 +181,7 @@ class TestDistanceAlgorithms(unittest.TestCase):
 					self.assertEqual(len(dists), nSources)
 					for distList in dists:
 						self.assertEqual(len(distList), nTargets)
-
+	
 	def testMultiTargetSTSP(self):
 		for g in self.genERGraphs():
 			source = nk.graphtools.randomNode(g)
@@ -198,6 +215,31 @@ class TestDistanceAlgorithms(unittest.TestCase):
 						self.assertEqual(pll.query(u, v), nk.none)
 					else:
 						self.assertEqual(pll.query(u, v), int(apsp.getDistance(u, v)))
+
+	def testSPSP(self):
+		nk.engineering.setSeed(1, True)
+		random.seed(1)
+		for directed in [True, False]:
+			for weighted in [True, False]:
+				g = nk.generators.ErdosRenyiGenerator(100, 0.15, directed).generate()
+				if weighted:
+					g = nk.graphtools.toWeighted(g)
+					g.forEdges(lambda u, v, ew, eid: g.setWeight(u, v, random.random()))
+				for nSources in [1, 10, 50]:
+					sources = set()
+					while len(sources) < nSources:
+						sources.add(nk.graphtools.randomNode(g))
+
+					spsp = nk.distance.SPSP(g, list(sources))
+					spsp.run()
+					spsp.getDistances()
+
+	def testWithNonConsecutiveNodeLabelsAPSP(self):
+		nk.engineering.setSeed(1, True)
+		random.seed(1)
+		g = nk.generators.ErdosRenyiGenerator(100, 0.15).generate()
+		g.removeNode(0)
+		nk.distance.APSP(g).run()	
 
 if __name__ == "__main__":
 	unittest.main()
