+++ networkit/test/test_generators.py
@@ -3,21 +3,16 @@ import unittest
 import networkit as nk
 
 class TestGraph(unittest.TestCase):
-    def testPubWebGenerator(self):
-        n = 100
-        gen = nk.generators.PubWebGenerator(n, 8, 0.1, 4)
+    
+    def testClusteredRandomGraphGenerator(self):
+        n, k = 100, 10
+        pIntra, pInter = 0.1, 0.05
+        gen = nk.generators.ClusteredRandomGraphGenerator(n, k, pIntra, pInter)
         G = gen.generate()
-        coords = gen.getCoordinates()
-
         self.assertEqual(G.numberOfNodes(), n)
-        self.assertEqual(len(coords), n)
-        self.assertTrue(all(len(pt) == 2 for pt in coords))
-
-        # check that all points are within the unit square and that they are non-empt
-        self.assertTrue(all(0.0 <= pt[0] <= 1 for pt in coords))
-        self.assertTrue(any(0.0 <  pt[0] <  1 for pt in coords))
-        self.assertTrue(all(0.0 <= pt[1] <= 1 for pt in coords))
-        self.assertTrue(any(0.0 <  pt[1] <  1 for pt in coords))
+        nCommunities = len(set(gen.getCommunities().getVector()))
+        self.assertGreater(nCommunities, 0)
+        self.assertLessEqual(nCommunities, k)
 
     def testDynamicPubWebGenerator(self):
         n = 100
@@ -40,26 +35,7 @@ class TestGraph(unittest.TestCase):
         newCoords = gen.getNewCoordinates()
         self.assertGreater(len(newCoords), 0)
         self.assertEqual(newCoords[0][0], n)
-
-    def testWattsStrogatzGenerator(self):
-        n = 12
-        nbrs = 5
-        G = nk.generators.WattsStrogatzGenerator(n, nbrs, 0.5).generate()
-
-        self.assertEqual(G.numberOfNodes(), n)
-        self.assertEqual(G.numberOfEdges(), n*nbrs)
-
-    def testClusteredRandomGraphGenerator(self):
-        n, k = 100, 10
-        pIntra, pInter = 0.1, 0.05
-        gen = nk.generators.ClusteredRandomGraphGenerator(n, k, pIntra, pInter)
-        G = gen.generate()
-        self.assertEqual(G.numberOfNodes(), n)
-        nCommunities = len(set(gen.getCommunities().getVector()))
-        self.assertGreater(nCommunities, 0)
-        self.assertLessEqual(nCommunities, k)
-
-
+    
     def testEdgeSwitchingMarkovChainGenerator(self):
         n = 100
         while True:
@@ -74,6 +50,32 @@ class TestGraph(unittest.TestCase):
                 self.assertEqual(G.degree(u), degseq[u])
 
             break
+    
+    def testPubWebGenerator(self):
+        n = 100
+        gen = nk.generators.PubWebGenerator(n, 8, 0.1, 4)
+        G = gen.generate()
+        coords = gen.getCoordinates()
+
+        self.assertEqual(G.numberOfNodes(), n)
+        self.assertEqual(len(coords), n)
+        self.assertTrue(all(len(pt) == 2 for pt in coords))
+
+        # check that all points are within the unit square and that they are non-empt
+        self.assertTrue(all(0.0 <= pt[0] <= 1 for pt in coords))
+        self.assertTrue(any(0.0 <  pt[0] <  1 for pt in coords))
+        self.assertTrue(all(0.0 <= pt[1] <= 1 for pt in coords))
+        self.assertTrue(any(0.0 <  pt[1] <  1 for pt in coords))
+
+    def testWattsStrogatzGenerator(self):
+        n = 12
+        nbrs = 5
+        G = nk.generators.WattsStrogatzGenerator(n, nbrs, 0.5).generate()
+
+        self.assertEqual(G.numberOfNodes(), n)
+        self.assertEqual(G.numberOfEdges(), n*nbrs)
+
+    
 
 if __name__ == "__main__":
     unittest.main()
