+++ networkit/test/test_matching.py
@@ -6,13 +6,13 @@ from copy import copy
 
 import networkit as nk
 
-class TestMatchingAlgorithms(unittest.TestCase):
+class TestMatching(unittest.TestCase):
 
 	def setUp(self):
 		self.g = nk.readGraph("input/PGPgiantcompo.graph", nk.Format.METIS)
 		self.gw = copy(self.g)
 		nk.graphtools.randomizeWeights(self.gw)
-
+	
 	def hasUnmatchedNeighbors(self, g, m):
 		for e in g.iterEdges():
 			if not m.isMatched(e[0]) and not m.isMatched(e[1]):
