+++ networkit/test/test_traversal.py
